♻️ refactor: useAuth hook 이동 - /application/hooks/domain -> /features/common/hooks
