--- src/application/hooks/domain/auth/index.ts
@@ -1,1 +1,0 @@
-export * from "./useAuth";
