--- src/application/hooks/domain/index.ts
@@ -1,1 +1,0 @@
-export * from "./auth";
