+++ src/features/common/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from "./useAuth";
 export * from "./useChannelIO";
 export * from "./useCollection";
 export * from "./useMoveMemeDetail";
