+++ src/pages/collect/index.tsx
@@ -1,12 +1,11 @@
 import { useDeferredValue } from "react";
 
-import { useAuth } from "@/application/hooks";
 import { BackButtonNavigation } from "@/common/components/Navigation";
 import { MemeListSkeleton } from "@/common/components/Skeleton";
 import { SSRSuspense } from "@/common/components/Suspense";
 import { useDebounce, useInput } from "@/common/hooks";
 import { Collection, SearchedCollection } from "@/features/collect/components";
-import { withAuth } from "@/features/common";
+import { useAuth, withAuth } from "@/features/common";
 import { SearchInput } from "@/features/search/components";
 
 const CollectPage = () => {
